+++ gDR/R/runDrugResponseProcessingPipeline.R
@@ -9,12 +9,14 @@
 #' Defaults to \code{mean(x, trim = 0.25)}.
 #' @param nested_keys character vector of column names to include in the data.frames in the assays of the resulting \code{SummarizedExperiment} object.
 #' Defaults to \code{c("Barcode", gDRutils::get_identifier("masked_tag"))}.
-#' @param override_controls named list containing defining factors in the treatments.
+#' @param override_untrt_controls named list containing defining factors in the treatments.
 #' Defaults to \code{NULL}.
 #' @param override_masked boolean indicating whether or not to override the masked wells
 #' in the averaging and include all wells. 
 #' Defaults to \code{FALSE}.
 #' @param ndigit_rounding integer indicating number of digits to round to in calculations.
+#' Defaults to \code{4}.
+#' @param n_point_cutoff integer of how many points should be considered the minimum required to try to fit a curve.
 #' Defaults to \code{4}.
 #' @param control_assay string containing the name of the assay representing the controls in the \code{se}.
 #' Defaults to \code{"Controls"}.
@@ -39,9 +41,10 @@ runDrugResponseProcessingPipeline2 <- function(df_,
                                                readout = "ReadoutValue",
                                                control_mean_fxn = function(x) {mean(x, trim = 0.25)},
                                                nested_keys = c("Barcode", gDRutils::get_identifier("masked_tag")),
-                                               override_controls = NULL,
+                                               override_untrt_controls = NULL,
                                                override_masked = FALSE,
                                                ndigit_rounding = 4,
+                                               n_point_cutoff = 4,
                                                control_assay = "Controls",
                                                raw_treated_assay = "RawTreated",
                                                normalized_assay = "Normalized",
@@ -53,7 +56,7 @@ runDrugResponseProcessingPipeline2 <- function(df_,
                    readout = readout, 
                    control_mean_fxn = control_mean_fxn, 
                    nested_keys = nested_keys, 
-                   override_controls = override_controls)
+                   override_untrt_controls = override_untrt_controls)
   se <- normalize_SE2(se = se, 
                       control_assay = control_assay, 
                       raw_treated_assay = raw_treated_assay, 
@@ -69,7 +72,7 @@ runDrugResponseProcessingPipeline2 <- function(df_,
                 averaged_assay = averaged_assay, 
                 ref_GR_assay = ref_GR_assay, 
                 metrics_assay = metrics_assay, 
-                ndigit_rounding = ndigit_rounding)
+                n_point_cutoff = n_point_cutoff)
   se
 }